backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
numpy
whoosh
pandas
matplotlib
colorama
pathlib
wheel
nbresult
ipdb
ipykernel
yapf
pygeohash
pytest
seaborn
scipy
scikit-learn
google-cloud-bigquery
google-cloud-storage
db-dtypes
pyarrow
tensorflow==2.10.0
